# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VisiteCRM API - SaaS CRM for tourism agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["string", "null"]
        - name: channel
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── TEMPLATES ────────────────────────────────────────────────────
  /message-templates:
    get:
      operationId: listMessageTemplates
      tags: [templates]
      summary: List message templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageTemplate"
    post:
      operationId: createMessageTemplate
      tags: [templates]
      summary: Create a message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
  /message-templates/{id}:
    patch:
      operationId: updateMessageTemplate
      tags: [templates]
      summary: Update a message template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageTemplateBody"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
    delete:
      operationId: deleteMessageTemplate
      tags: [templates]
      summary: Delete a message template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── AUTOMATIONS ──────────────────────────────────────────────────
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List automations
      responses:
        "200":
          description: Automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create an automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    patch:
      operationId: updateAutomation
      tags: [automations]
      summary: Update an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /automations/{id}/toggle:
    patch:
      operationId: toggleAutomation
      tags: [automations]
      summary: Toggle automation active state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"

  # ─── SUPPLIERS ────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ─── DASHBOARD ─────────────────────────────────────────────────
...
    Message:
      type: object
      properties:
        id:
          type: string
        toClientId:
          type: ["string", "null"]
        channel:
          type: string
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
        status:
          type: string
        sentAt:
          type: string
        deliveredAt:
          type: ["string", "null"]
        readAt:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
      required: [id, channel, content, status, sentAt]
    SendMessageBody:
      type: object
      properties:
        toClientId:
          type: string
        channel:
          type: string
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
      required: [toClientId, channel, content]
    MessageTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        channel:
          type: string
        subject:
          type: ["string", "null"]
        content:
          type: string
        variables:
          type: array
          items:
            type: string
        category:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, channel, content, variables, createdAt]
    CreateMessageTemplateBody:
      type: object
      properties:
        name:
          type: string
        channel:
          type: string
        subject:
          type: ["string", "null"]
        content:
          type: string
        variables:
          type: array
          items:
            type: string
        category:
          type: ["string", "null"]
      required: [name, channel, content]
    UpdateMessageTemplateBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        category:
          type: ["string", "null"]

    # ─── AUTOMATION ────────────────────────────────────────────────
    Automation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        triggerType:
          type: string
        triggerConfig:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
        executionsCount:
          type: integer
        lastExecutedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, triggerType, triggerConfig, isActive, executionsCount, createdAt]
    CreateAutomationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        triggerType:
          type: string
        triggerConfig:
          type: object
          additionalProperties: true
        conditions:
          type: ["object", "null"]
          additionalProperties: true
      required: [name, triggerType, triggerConfig]
    UpdateAutomationBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        triggerConfig:
          type: ["object", "null"]
          additionalProperties: true

    # ─── SUPPLIER ──────────────────────────────────────────────────
...